Refactor: Split invitation service into modules
This commit refactors the invitation service to improve modularity and maintainability by splitting it into smaller, more focused files.
